feat(FileShared): Add a BasePreviewSliderFragment class
