+++ app/src/main/java/com/infomaniak/drive/ui/fileList/preview/PreviewFragment.kt
@@ -22,9 +22,9 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
+import androidx.navigation.fragment.navArgs
 import androidx.navigation.navGraphViewModels
 import com.infomaniak.drive.R
 import com.infomaniak.drive.data.cache.FileController
@@ -34,10 +34,11 @@ import io.sentry.Sentry
 
 open class PreviewFragment : Fragment() {
 
-    protected lateinit var file: File
     private val mainViewModel: MainViewModel by activityViewModels()
     private val previewViewModel: PreviewViewModel by viewModels()
-    protected val previewSliderViewModel: PreviewSliderViewModel by navGraphViewModels(R.id.previewSliderFragment)
+    protected val previewSliderViewModel: PreviewSliderViewModel by viewModels(ownerProducer = ::requireParentFragment)
+
+    protected lateinit var file: File
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         if (previewViewModel.currentFile == null) {
@@ -54,29 +55,23 @@ open class PreviewFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
     }
 
-    private fun getCurrentFile(fileId: Int): File? {
-        return try {
-            FileController.getFileById(fileId, previewSliderViewModel.userDrive) ?: mainViewModel.currentPreviewFileList[fileId]
-        } catch (exception: Exception) {
-            exception.printStackTrace()
-            Sentry.withScope { scope ->
-                val backStackEntry = findNavController().currentBackStackEntry
-                val previousName = findNavController().previousBackStackEntry?.destination?.displayName
-                scope.setExtra("destination", "${backStackEntry?.destination?.displayName}")
-                scope.setExtra("destination lifecycle", "${backStackEntry?.lifecycle?.currentState}")
-                scope.setExtra("previous", previousName ?: "")
-                scope.setExtra("exception", exception.stackTraceToString())
-                Sentry.captureException(exception)
-            }
-            null
+    private fun getCurrentFile(fileId: Int): File? = runCatching {
+        FileController.getFileById(fileId, previewSliderViewModel.userDrive) ?: mainViewModel.currentPreviewFileList[fileId]
+    }.getOrElse { exception ->
+        exception.printStackTrace()
+        Sentry.withScope { scope ->
+            val backStackEntry = findNavController().currentBackStackEntry
+            val previousName = findNavController().previousBackStackEntry?.destination?.displayName
+            scope.setExtra("destination", "${backStackEntry?.destination?.displayName}")
+            scope.setExtra("destination lifecycle", "${backStackEntry?.lifecycle?.currentState}")
+            scope.setExtra("previous", previousName ?: "")
+            scope.setExtra("exception", exception.stackTraceToString())
+            Sentry.captureException(exception)
         }
+        null
     }
 
     protected fun noCurrentFile() = previewViewModel.currentFile == null
-
-    protected class PreviewViewModel : ViewModel() {
-        var currentFile: File? = null
-    }
 
     companion object {
         const val FILE_ID_TAG = "file_id_tag"
